moved action and util into navbar module
