+++ frontend/core/navbar/subcomponents/index.js
@@ -1,3 +1,4 @@
 export * from './signedin_dropdown';
 export * from './pool_dropdown';
 export * from './signedout_dropdown';
+export * from './confirm_form';
