+++ frontend/core/navbar/utils.js
